group the governance commands
